# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Documentation of Server to add/delete/update/get employee and department detail",
        "description": "Employee Department Management"
    },
    "servers": [],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/employee": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get all employee in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Create a new employee in system",
                "requestBody": {
                    "description": "Employee Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Employee"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "422": {
                        "description": "Failed. Invalid department id."
                    }
                }
            }
        },
        "/employee/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the employee that we want to match",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get employee with given ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Employee with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    },
                    "404": {
                        "description": "Failed. Employee not found."
                    }
                }
            },
            "put": {
                "summary": "Update employee with given ID",
                "tags": [
                    "Employee"
                ],
                "requestBody": {
                    "description": "Employee Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Employee"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Employee with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    },
                    "422": {
                        "description": "Failed. Unprocessable data."
                    }
                }
            },
            "delete": {
                "summary": "delete employee with given ID",
                "tags": [
                    "Employee"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Delete Employee with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "422": {
                        "description": "Failed. Employee not found."
                    }
                }
            }
        },
        "/employee/sortby/{key}": {
            "parameters": [
                {
                    "name": "key",
                    "in": "path",
                    "required": true,
                    "description": "column name of employee table",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get sorted value of any column(e_name,age,dept,e_id)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    },
                    "422": {
                        "description": "Failed. Invalid column name"
                    }
                }
            }
        },
        "/employee/searchby/{key}": {
            "parameters": [
                {
                    "name": "key",
                    "in": "path",
                    "required": true,
                    "description": "Search data",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Search any keyword",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    },
                    "404": {
                        "description": "Failed. No rows found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "Employee": {
            "type": "object",
            "properties": {
                "e_name": {
                    "type": "string"
                },
                "dept": {
                    "type": "integer"
                },
                "age": {
                    "type": "integer"
                }
            }
        },
...
    }
}